Decode routing selectors through one shared app-side helper
The fail-closed workspace flag was set by the coordinator's decoder but not
by the two app-side request decoders, so `surface.read_text` on the
socket-worker lane (and the remote-tmux paths) still treated an explicit
unresolvable `workspace_id` as omitted and read the focused workspace.

Both now go through `TerminalController.v2RoutingSelectors(_:)`, leaving
exactly two decoders in the codebase (this one and the coordinator's), so a
selector cannot be wired into one entrypoint and forgotten in another.
