+++ Sources/TerminalController+RemoteTmux.swift
@@ -202,16 +202,7 @@ extension TerminalController {
     }
 
     nonisolated func remoteTmuxRouting(from params: [String: Any]) -> ControlRoutingSelectors {
-        ControlRoutingSelectors(
-            hasWindowIDParam: v2HasNonNullParam(params, "window_id"),
-            windowID: v2UUID(params, "window_id"),
-            groupID: v2UUID(params, "group_id"),
-            workspaceID: v2UUID(params, "workspace_id"),
-            surfaceID: v2UUID(params, "surface_id")
-                ?? v2UUID(params, "terminal_id")
-                ?? v2UUID(params, "tab_id"),
-            paneID: v2UUID(params, "pane_id")
-        )
+        v2RoutingSelectors(params)
     }
 
     private nonisolated static func remoteTmuxActivate(from params: [String: Any]) -> Bool {
